# Conflicts:
#	appbase/src/main/res/values-sw240dp/dimens.xml
#	appbase/src/main/res/values-sw360dp/dimens.xml
#	appbase/src/main/res/values-sw480dp/dimens.xml
#	appbase/src/main/res/values-sw600dp/dimens.xml
#	appbase/src/main/res/values-sw720dp/dimens.xml
#	appbase/src/main/res/values-sw840dp/dimens.xml
#	appbase/src/main/res/values-sw960dp/dimens.xml

+++ appbase/src/main/java/tv/ismar/app/core/PageIntent.java
@@ -128,7 +128,7 @@ public class PageIntent implements PageIntentInterface {
     }
 
     @Override
-    public void toPackageList(Context context, String source, long pk) {
+    public void toPackageList(Context context, String source, int pk) {
         Intent intent=new Intent();
         intent.setAction("tv.ismar.daisy.packagelist");
         intent.putExtra("pk",pk);
